Internals

alfa.Crystal

alfa.CrystalOperator

alfa.Lattice

alfa.Multiplier

alfa.OperatorComposition

Base.:*

Base.:*

Base.:+

Base.:/

Base.getindex

Base.getproperty

Base.lcm

Base.push!

Base.size

alfa.CheckIfNormal

alfa.CleanUp!

alfa.CrystalOperatorCopyLowerTriangle

alfa.CrystalOperatorCopyWithMultipliers

alfa.ElementsInQuotientSpace

alfa.ShiftIntoStandardCell

alfa.eigen

alfa.eigvals

alfa.eigvals

alfa.eigvals_df

alfa.find_multiplier

alfa.hnf

alfa.lll

alfa.normalize

alfa.normalize

alfa.snf_with_transform

alfa.symbol

alfa.wrtLattice

alfa.wrtLattice

alfa.wrtSameLatticeAndNormalize

alfa.Crystal — Method

Crystal{N,T}(L = nothing, Domain = nothing, Codomain = nothing) where {N,T<:Union{Float64, Rational}}

Constructs a Crystal{N,T} which respresents the domain and codomain of a CrystalOperator. It consists of a L::Lattice{N,T} and the structure elements Domain::Vector{SVector{N,T}} and Codomain::Vector{SVector{N,T}}.

This struct describes the set

\[Aℤ^N+s = \left\{ ∑_{i=0}^n z_j ⋅ a_j + (s_1,s_2,\ldots,s_m) : z_j ∈ ℤ \right\},\]

where $s=(s_1,s_2,\ldots,s_m) ∈ \{\text{Domain,Codomain}\}$ is the structure element, $A$ the lattice basis L.A.

In case of L==nothing, the identity I of size N is used.

In case of Domain==nothing, the origin $0 \in \mathbb{R}^N$ is used.

In case Codomain==nothing, Domain is used.

Example

julia> using alfa

julia> alfa.Crystal{2,Float64}()
Lattice Basis: alfa.Lattice{2,Float64}([1.0 0.0; 0.0 1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]

alfa.CrystalOperator — Method

CrystalOperator(
    C::Crystal{N,T},
    J::UniformScaling,
    _CompatibilityCheckOnly = false,
) where {N,T}
CrystalOperator{N,T}(
    C = nothing,
    M = nothing,
    _CompatibilityCheckOnly = false,
) where {N,T<:Union{Float64,Rational}}

Constructs a translationally invariant C::CrystalOperators{N,T}

\[C : \mathcal{L}(C.L^\text{Domain}) \rightarrow \mathcal{L}(C.L^\text{Codomain})\]

The actual function definition $(Cf)(x)$ is saved within C.M::SortedSet{Multiplier}:

\[(Cf)(x) = \sum_{y \in C.M} y.\text{mat} ⋅ f(x+y.\text{pos}) \quad \forall x \in (C.L.A)\mathbb{Z}^N\]

Example

julia> using alfa

julia> using LinearAlgebra

julia> alfa.CrystalOperator{2,Float64}()
Lattice Basis: alfa.Lattice{2,Float64}([1.0 0.0; 0.0 1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
Multiplier: 0-element Array{alfa.Multiplier,1}

julia> alfa.CrystalOperator(alfa.Crystal{2,Float64}(),3*I)
Lattice Basis: alfa.Lattice{2,Float64}([1.0 0.0; 0.0 1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
Multiplier: 1-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{2}([0, 0], [3])

alfa.Lattice — Method

Lattice{N,T}(A=nothing)

Construct a Lattice{N,T} with basis $A$, i.e., it represents the set

\[Aℤ^N = \left\{ ∑_{i=0}^n z_j ⋅ a_j : z_j ∈ ℤ \right\},\]

where $a_j$ denotes the $j$th column of $A$. The matrix $A$ must be square and nonsingular:

$A$ is called the lattice basis.

$a_j$ are the primite vectors of the lattice.

T<:Union{Float64,Rational} represents the datatype of the entries of the basis $a_{ij}$.

In case of A==nothing, the identity I of size N is used.

See Definition 2.1 in [1].

Example

julia> using alfa

julia> L = alfa.Lattice{3,Float64}()
alfa.Lattice{3,Float64}([1.0 0.0 0.0; 0.0 1.0 0.0; 0.0 0.0 1.0])

julia> L = alfa.Lattice{2,Rational{BigInt}}()
alfa.Lattice{2,Rational{BigInt}}(Rational{BigInt}[1//1 0//1; 0//1 1//1])

julia> L = alfa.Lattice{2,Rational{BigInt}}([1 2; 3 4])
alfa.Lattice{2,Rational{BigInt}}(Rational{BigInt}[1//1 2//1; 3//1 4//1])

julia> L = alfa.Lattice{2,Rational{BigInt}}([1 2; 1 2])
ERROR: AssertionError: Basis must be nonsingular

alfa.Multiplier — Method

Multiplier{N}(pos = nothing, mat = nothing) where {N}
Multiplier(pos = nothing, mat = nothing)

Constructs a multiplication matrix as part of a CrystalOperator. The position is given in fractional coordinates and thus (converted to) integral.

Example

julia> using alfa

julia> alfa.Multiplier([0 0], [1 2 3; 4 5 6])
Position: 2-element StaticArrays.MArray{Tuple{2},Int64,1,2} with indices SOneTo(2):
 0
 0
Multiplier: 2×3 Array{Int64,2}:
 1  2  3
 4  5  6

alfa.OperatorComposition — Method

OperatorComposition(f::Expr)

Constructs a struct OperatorComposition from an expression. All CrystalOperators are rewritten with respect to a single translationally invariance and normalized. Furthermore, it is checked if the expression is well-defined.

Example

julia> using alfa

julia> L = alfa.gallery.Laplace(N=1);

julia> f = :(inv($L)*$L);

julia> oc = alfa.OperatorComposition(f);

julia> alfa.symbol(oc,[randn(1)]) ≈ [1] # as f is the identitity, the fourier transform is 1 for all frequencies.
true

Base.:* — Method

Base.:*(A::CrystalOperator{N,T}, B::CrystalOperator{N,T}) where {N,T}

If there is a least common multiple translationally invariance of A and B, then both operators A and B are rewritten wrt this translationally invariance A2 and B2. After that, it is checked if domain and codomain are compatible. If thats the case, the crystaloperator A2\cdot B2 is constructed.

\[A2\cdot B2 : \mathcal{L}(B2.C.L^\text{Domain}) \rightarrow \mathcal{L}(A2.C.L^\text{Codomain})\]

with

\[(A2\cdot B2\cdot f)(x) = \sum_{a \in A2.M, b \in B2.M} a.\text{mat}\cdot b.\text{mat} ⋅ f(x+a.\text{pos}+b.\text{pos}) \quad \forall x \in (A2.C.L.A)\mathbb{Z}^N.\]

Example

julia> using alfa

julia> A = alfa.gallery.Laplace(N=1)
Lattice Basis: alfa.Lattice{1,Float64}([1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Multiplier: 3-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{1}([-1], [1.0])
 alfa.Multiplier{1}([0], [-2.0])
 alfa.Multiplier{1}([1], [1.0])

julia> R = alfa.gallery.fw_restriction(N=1)
Lattice Basis: alfa.Lattice{1,Float64}([2.0])
Domain: 2-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
 [1.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Multiplier: 2-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{1}([-1], Rational{Int64}[0//1 1//2])
 alfa.Multiplier{1}([0], Rational{Int64}[1//1 1//2])

julia> R*A
Lattice Basis: alfa.Lattice{1,Float64}([2.0])
Domain: 2-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
 [1.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Multiplier: 3-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{1}([-1], [0.5 0.0])
 alfa.Multiplier{1}([0], [-1.0 0.0])
 alfa.Multiplier{1}([1], [0.5 0.0])

Base.:* — Method

Base.:*(b::T, A::CrystalOperator) where {T<:Number}
Base.:*(A::CrystalOperator{N,T}, b::S) where {N,T,S<:Number}

Constructs a translationally invariant C::CrystalOperators{N,T}

\[b\cdot C : \mathcal{L}(C.L^\text{Domain}) \rightarrow \mathcal{L}(C.L^\text{Codomain})\]

with

\[(b\cdot C\cdot f)(x) = b\cdot \sum_{y \in C.M} y.\text{mat} ⋅ f(x+y.\text{pos}) \quad \forall x \in (C.L.A)\mathbb{Z}^N\]

Example

julia> using alfa

julia> L = alfa.gallery.Laplace(N=1)
Lattice Basis: alfa.Lattice{1,Float64}([1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Multiplier: 3-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{1}([-1], [1.0])
 alfa.Multiplier{1}([0], [-2.0])
 alfa.Multiplier{1}([1], [1.0])

julia> 2L
Lattice Basis: alfa.Lattice{1,Float64}([1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Multiplier: 3-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{1}([-1], [2.0])
 alfa.Multiplier{1}([0], [-4.0])
 alfa.Multiplier{1}([1], [2.0])

Base.:+ — Method

Base.:+(A::CrystalOperator{N,T}, B::CrystalOperator{N,T}) where {N,T}

If there is a least common multiple translationally invariance of A and B, then both operators A and B are rewritten wrt this translationally invariance A2 and B2. After that, it is checked if domain and codomain are compatible. If thats the case, the crystaloperator A2+t B2 is constructed.

\[A2+B2 : \mathcal{L}(A2.C.L^\text{Domain}) \rightarrow \mathcal{L}(A2.C.L^\text{Codomain})\]

with

\[(A2+B2\cdot f)(x) = (A2\cdot f)(x) + (B2\cdot f)(x) \quad \forall x \in (A2.C.L.A)\mathbb{Z}^N.\]

Example

julia> using alfa

julia> A = alfa.gallery.Laplace(N=1)
Lattice Basis: alfa.Lattice{1,Float64}([1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Multiplier: 3-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{1}([-1], [1.0])
 alfa.Multiplier{1}([0], [-2.0])
 alfa.Multiplier{1}([1], [1.0])

julia> A+A
Lattice Basis: alfa.Lattice{1,Float64}([1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Multiplier: 3-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{1}([-1], [2.0])
 alfa.Multiplier{1}([0], [-4.0])
 alfa.Multiplier{1}([1], [2.0])

Base.:/ — Method

Base.:/(A::CrystalOperator, b::T) where {T<:Number}

Constructs a translationally invariant C::CrystalOperators{N,T}

\[C/b : \mathcal{L}(C.L^\text{Domain}) \rightarrow \mathcal{L}(C.L^\text{Codomain})\]

with

\[(C/b\cdot f)(x) = \frac{1}{b} \sum_{y \in C.M} y.\text{mat} ⋅ f(x+y.\text{pos}) \quad \forall x \in (C.L.A)\mathbb{Z}^N\]

Example

julia> using alfa

julia> L = alfa.gallery.Laplace(N=1)
Lattice Basis: alfa.Lattice{1,Float64}([1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Multiplier: 3-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{1}([-1], [1.0])
 alfa.Multiplier{1}([0], [-2.0])
 alfa.Multiplier{1}([1], [1.0])

julia> L/2
Lattice Basis: alfa.Lattice{1,Float64}([1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Multiplier: 3-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{1}([-1], [0.5])
 alfa.Multiplier{1}([0], [-1.0])
 alfa.Multiplier{1}([1], [0.5])

Base.getindex — Method

Base.getindex(L::Lattice, y...)

simply wrapping getindex(L.A, y...).

Example

julia> using alfa

julia> L = alfa.Lattice{2,Float64}([1 2; 3 4])
alfa.Lattice{2,Float64}([1.0 2.0; 3.0 4.0])
julia> L[1,2] == L.A[1,2]
true

Base.getproperty — Method

Base.getproperty(L::Lattice, sym::Symbol)

Get properties of Lattice{N,T}. Let $A=$L.A, i.e., $A ∈ T^{N × N}$, then

L.n and L.dim return the dimension N.

L.iA returns $A^{-1}$, i.e., the inverse of the lattice basis

L.dA returns $A^{-T}$, i.e., the basis of the dual lattice.

Example

julia> using alfa

julia> L = alfa.Lattice{2,Float64}([1 2; 3 4])
alfa.Lattice{2,Float64}([1.0 2.0; 3.0 4.0])
julia> L.n == L.dim == typeof(L).parameters[1] == 2
true
julia> L[2,1]
3.0
julia> L.iA == inv(L.A)
true
julia> L.dA == transpose(inv(L.A))
true

Base.lcm — Method

Base.lcm(A::MArray{X,T},B::MArray{X,T}...)
Base.lcm(A::MArray{X,T}, B::MArray{X,T}...) where {X,T<:Rational}

Base.lcm(A::Lattice{N,T}, B::Lattice{N,T}...) where {N,T}
Base.lcm(A::MArray{X,T}, B::MArray{X,T}...) where {X,T<:Rational}

Returns the least common multiple of $A$ and $B$ (or more), i.e. a sub-lattice $C$, that is $C ⊂ A$ and $C ⊂ B$ with $|\det(C)|$ as small as possible.

See Algorithm B.7, Theorem 5.1 in [1].

Example

julia> using alfa

julia> L = alfa.Lattice{2,Float64}([1 2; 3 4])
alfa.Lattice{2,Float64}([1.0 2.0; 3.0 4.0])

julia> alfa.lcm(L) == L
true

julia> alfa.lcm(L.A) == L.A
true

julia> alfa.lcm(alfa.Lattice{1,Float64}([2]), alfa.Lattice{1,Float64}([3]), alfa.Lattice{1,Float64}([6]))
alfa.Lattice{1,Float64}([6.0])

julia> alfa.lcm(alfa.Lattice{2,Rational{BigInt}}([1 1; -1 1]), alfa.Lattice{2,Rational{BigInt}}([1 2; 2 1]))
alfa.Lattice{2,Rational{BigInt}}(Rational{BigInt}[-3//1 1//1; -3//1 -1//1])

Base.push! — Function

Base.push!(S::CrystalOperator, m::Multiplier, add_to_existing = false)

Adds a multiplier to S.M. If there is a multiplier m2 in S.M with m2.pos == m.pos, then

m2 is replaced if add_to_existing == false

m2.mat+m.mat is the new multiplier at m.pos if add_to_existing == true.

Example

julia> using alfa

julia> S = alfa.gallery.Laplace()
Lattice Basis: alfa.Lattice{2,Float64}([1.0 0.0; 0.0 1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
Multiplier: 5-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{2}([-1, 0], [1.0])
 alfa.Multiplier{2}([0, -1], [1.0])
 alfa.Multiplier{2}([0, 0], [-4.0])
 alfa.Multiplier{2}([0, 1], [1.0])
 alfa.Multiplier{2}([1, 0], [1.0])

julia> m = alfa.Multiplier{2}([0, 0], [1])
Position: 2-element StaticArrays.MArray{Tuple{2},Int64,1,2} with indices SOneTo(2):
 0
 0
Multiplier: 1×1 Array{Int64,2}:
 1

julia> push!(S,m)
Lattice Basis: alfa.Lattice{2,Float64}([1.0 0.0; 0.0 1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
Multiplier: 5-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{2}([-1, 0], [1.0])
 alfa.Multiplier{2}([0, -1], [1.0])
 alfa.Multiplier{2}([0, 0], [1])
 alfa.Multiplier{2}([0, 1], [1.0])
 alfa.Multiplier{2}([1, 0], [1.0])

julia> push!(S,m, true)
Lattice Basis: alfa.Lattice{2,Float64}([1.0 0.0; 0.0 1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
Multiplier: 5-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{2}([-1, 0], [1.0])
 alfa.Multiplier{2}([0, -1], [1.0])
 alfa.Multiplier{2}([0, 0], [2])
 alfa.Multiplier{2}([0, 1], [1.0])
 alfa.Multiplier{2}([1, 0], [1.0])

Base.size — Method

Base.size(L::Lattice)

Returns the dimension of the lattice basis: size(L.A,1).

Example

julia> using alfa

julia> L = alfa.Lattice{2,Float64}([1 2; 3 4])
alfa.Lattice{2,Float64}([1.0 2.0; 3.0 4.0])
julia> size(L) == size(L.A) == (2,2)
true

alfa.CheckIfNormal — Method

CheckIfNormal(s, A)

Checks if s is found within the primtive cell of A, i.e., $A^{-1}s_j ∈ [0,1)^N$ and if s is sorted lexicographically.

See Definition 5.4 in [1].

Example

julia> using alfa

julia> using LinearAlgebra

julia> L = alfa.Lattice{2,Float64}([1 0; 0 1]);

julia> s = [[-1/2,0], [1/4,2]]
2-element Array{Array{Float64,1},1}:
 [-0.5, 0.0]
 [0.25, 2.0]

julia>  alfa.CheckIfNormal(s,L)
false

julia> t = [[1/4,0], [1/2,0]]
2-element Array{Array{Float64,1},1}:
 [0.25, 0.0]
 [0.5, 0.0]

julia> alfa.CheckIfNormal(t,L)
true

alfa.CleanUp! — Method

CleanUp!(S::CrystalOperator)

Removes all multipliers m with norm(m.mat) == 0.

Example

julia> using alfa

julia> using LinearAlgebra

julia> S = alfa.CrystalOperator(alfa.Crystal{2,Float64}(),0*I)
Lattice Basis: alfa.Lattice{2,Float64}([1.0 0.0; 0.0 1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
Multiplier: 1-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{2}([0, 0], [0])

julia> alfa.CleanUp!(S)

julia> S
Lattice Basis: alfa.Lattice{2,Float64}([1.0 0.0; 0.0 1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
Multiplier: 0-element Array{alfa.Multiplier,1}

alfa.CrystalOperatorCopyLowerTriangle — Method

CrystalOperatorCopyLowerTriangle(S::CrystalOperator{N,T}; idx = nothing, omega=1.0) where {N,T}

Returns the CrystalOperator G consisting of the central multiplier of S: mG^{(0)} = p@mL^{(pos)}@p if pos ≦ 0, m_G^{(y)} = 0 if pos > 0.

with P[i,j] = 1 if i=j in idx,
    = 0 else.

alfa.CrystalOperatorCopyWithMultipliers — Method

CrystalOperatorCopyWithMultipliers(S::CrystalOperator{N,T}; pos = nothing, idx = nothing) where {N,T}

Returns the CrystalOperator G consisting of the central multiplier of S: mG^{(0)} = p@mL^{(pos)}@p, m_G^{(y)} = 0 if y != pos,

with P[i,j] = 1 if i=j in idx,
    = 0 else.

alfa.ElementsInQuotientSpace — Method

ElementsInQuotientSpace(
    A::Union{Matrix{T},MMatrix{N,N,T}},
    B::Union{Matrix{T},MMatrix{N,N,T}};
    return_diag_hnf::Bool = false,
    return_fractional::Bool = false,
) where {N,T}

Returns all lattice points of the lattice generated by $A$ found in the primitive cell of B, i.e.,

\[T_{A,B}=\{x : x ∈ A \mathbb{Z}^N \cap B[0,1)^N \}\]

See Algorithm B.3, Theorem 2.3 in [1].

Example

julia> using alfa

julia> A = alfa.Lattice{2,Float64}([1 0; 0 1]);

julia> B = alfa.Lattice{2,Float64}([2 0; 0 2]);

julia> alfa.ElementsInQuotientSpace(A.A,B.A)
4-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
 [1.0, 0.0]
 [0.0, 1.0]
 [1.0, 1.0]

alfa.ShiftIntoStandardCell — Method

ShiftIntoStandardCell(s, A::Union{Matrix{T},MMatrix{N,N,T}}) where {N,T}
ShiftIntoStandardCell(s, A::Union{Matrix{T},MMatrix{N,N,T}}) where {N,T<:Rational}
ShiftIntoStandardCell(s, A::Lattice)

Shifts all elements s[i] into the primitive cell $A[0,1)^N$ and sort the entries lexicographically. The function returns t, y, p, such that

\[t_j + A y_j = s_{p(j)}\]

t is the shifted vector s, i.e., $A^{-1}t_j ∈ [0,1)^N$ for all j.

p is the permutation with respect to the input s.

y corresponds to the shift in fractional coordinates.

See Definition 2.2, 5.2 in [1].

Example

julia> using alfa

julia> using LinearAlgebra

julia> L = alfa.Lattice{2,Float64}([1 0; 0 1]);

julia> s = [[-1/2,0], [1/4,2]]
2-element Array{Array{Float64,1},1}:
 [-0.5, 0.0]
 [0.25, 2.0]

julia> (t, y, p) = alfa.ShiftIntoStandardCell(s,L);

julia> t
2-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.25, 0.0]
 [0.5, 0.0]

julia> y
2-element Array{StaticArrays.MArray{Tuple{2},Float64,1,2},1}:
 [0.0, 2.0]
 [-1.0, 0.0]

julia> p
2-element Array{Int64,1}:
 2
 1

julia> [t[j] + L.A*y[j] - s[p[j]] for j in [1,2]]
2-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
 [0.0, 0.0]

alfa.eigen — Method

eigen(S::X, k; by = abs) where {X<:Union{CrystalOperator,OperatorComposition}}

Computes the eigenvalues and eigenvectors of the symbol of a CrystalOperator/OperatorComposition wrt frequency k.

The eigenvalues are sorted by by.

alfa.eigvals — Method

eigvals(
    S::X,
    k::T;
    by = abs,
) where {
    T<:Union{AbstractVector,Tuple},
    X<:Union{CrystalOperator,OperatorComposition},
}

Computes the eigenvalues of the symbol of a CrystalOperator/OperatorComposition wrt frequency k.

The eigenvalues are sorted by by.

alfa.eigvals — Method

eigvals(
    S::X;
    N = 20,
    by = abs,
    unique = false,
    digits = 5,
) where {X<:Union{CrystalOperator,OperatorComposition}}

Computes the eigenvalues of the symbol of a CrystalOperator/OperatorComposition. The Frequency space is divided into N^dim equidistant (unique) points frequencies k.

alfa.eigvals_df — Method

eigvals_df(S::X; N = 20, by = abs) where {X<:Union{CrystalOperator,OperatorComposition}}

Returns a dataframe with the eigenvalues of the symbol of a CrystalOperator/OperatorComposition. The Frequency space is divided into N^dim equidistant (unique) points frequencies k.

alfa.find_multiplier — Method

find_multiplier(S::CrystalOperator, pos)

Returns the multiplier at pos if existent (in S.M).

Example

julia> using alfa

julia> S = alfa.gallery.Laplace()
Lattice Basis: alfa.Lattice{2,Float64}([1.0 0.0; 0.0 1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
Multiplier: 5-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{2}([-1, 0], [1.0])
 alfa.Multiplier{2}([0, -1], [1.0])
 alfa.Multiplier{2}([0, 0], [-4.0])
 alfa.Multiplier{2}([0, 1], [1.0])
 alfa.Multiplier{2}([1, 0], [1.0])

julia> alfa.find_multiplier(S, [0, 0])
Position: 2-element StaticArrays.MArray{Tuple{2},Int64,1,2} with indices SOneTo(2):
 0
 0
Multiplier: 1×1 Array{Float64,2}:
 -4.0

alfa.hnf — Method

hnf(mat::MMatrix{M,N}) where {M, N}
hnf(mat::Matrix)

Wrapper of Nemo.hnf. Input is converted to BigInt. Returns H = mat*U, s.t. H is in Hermite Normal Form and U is unimodular.

Example

julia> using alfa # hide

julia> using LinearAlgebra # hide

julia> mat = rand(1:1000, 2, 2);

julia> H = alfa.hnf(mat);

julia> norm(LinearAlgebra.tril(H) - H) ≈ 0
true
julia> round(abs(det(inv(mat)*H)), digits=5)
1.0

alfa.lll — Method

lll(mat::MMatrix{M,N}) where {M, N}
lll(mat::Matrix)

Wrapper of Nemo.lll. Input is converted to BigInt. Applies the LLL-Algorithm to the input mat. Computes output L, such that mat*T=L for some unimodular T.

Example

julia> using alfa # hide

julia> using LinearAlgebra # hide

julia> mat = rand(1:1000, 2, 2);

julia> L = alfa.lll(mat);

julia> round(abs(det(inv(mat)*L)), digits=5)
1.0

alfa.normalize — Method

normalize(S::CrystalOperator{N,T}) where {N,T}

Normalizes the crystaloperator, i.e., returns an crystaloperator isomorphic to S where all structure elements are shifted in the standard cell $S.C.L.A\cdot[0,1)^N$ and sorted lexicographically.

Example

julia> using alfa

julia> S = alfa.CrystalOperator{1,Float64}(alfa.Crystal{1,Float64}([1], [-.5], [1.5]))
Lattice Basis: alfa.Lattice{1,Float64}([1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [-0.5]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [1.5]
Multiplier: 0-element Array{alfa.Multiplier,1}

julia> push!(S, alfa.Multiplier([0], [-2]))
Lattice Basis: alfa.Lattice{1,Float64}([1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [-0.5]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [1.5]
Multiplier: 1-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{1}([0], [-2])

julia> alfa.normalize(S)
Lattice Basis: alfa.Lattice{1,Float64}([1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.5]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.5]
Multiplier: 1-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{1}([-2], [-2])

alfa.normalize — Method

normalize(C::Crystal{N,T}) where {N,T}

Normalizes the crystal, i.e., shifts the structure elements Domain and Codomain into the standard primitive cell.

Example

julia> using alfa

julia> C = alfa.Crystal{1,Float64}([2], [[-1],[-.5],[-1.5]], [[0]])
Lattice Basis: alfa.Lattice{1,Float64}([2.0])
Domain: 3-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [-1.0]
 [-0.5]
 [-1.5]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]

julia> alfa.normalize(C)
Lattice Basis: alfa.Lattice{1,Float64}([2.0])
Domain: 3-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.5]
 [1.0]
 [1.5]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]

alfa.snf_with_transform — Method

snf_with_transform(L::Lattice)
snf_with_transform(mat::MMatrix{M,N}) where {M,N}
snf_with_transform(mat::Matrix)

Wrapper of Nemo.snfwithtransform. Input is converted to BigInt. Returns (S,U,V) such that UmatV = S, where S is the Smith normal form of mat.

Example

julia> using alfa # hide

julia> using LinearAlgebra # hide

julia> mat = rand(1:10, 10, 10);

julia> (S,U,V) = alfa.snf_with_transform(mat);

julia> norm(U*mat*V - S) ≈ 0
true

julia> abs(det(U)) ≈ abs(det(V)) ≈ 1
true

julia> norm(diagm(diag(S)) - S) ≈ 0
true

alfa.symbol — Method

symbol(S::CrystalOperator, k; π = π)
symbol(O::OperatorComposition, k; π = π)

Returns the symbol of the CrystalOperator/OperatorComposition for a given frequency/wavevector k.

Example

julia> using alfa

julia> L = alfa.gallery.Laplace(N=2);

julia> oc = alfa.OperatorComposition(:(3*$L));

julia> alfa.symbol(oc,[0.5, 0.5]) ≈ [-24] # as f is the identitity, the fourier transform is 1 for all frequencies.
true

alfa.wrtLattice — Method

wrtLattice(
    S::CrystalOperator{N,T},
    A,
    _CompatibilityCheckOnly = false,
    normalize = true
) where {N,T}

Rewrites the crystaloperator S wit hrespect to the translationally invariance A.

Example

julia> using alfa

julia> L = alfa.gallery.Laplace(N=1)
Lattice Basis: alfa.Lattice{1,Float64}([1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Multiplier: 3-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{1}([-1], [1.0])
 alfa.Multiplier{1}([0], [-2.0])
 alfa.Multiplier{1}([1], [1.0])

julia> alfa.wrtLattice(L, L.C.L.A*2)
Lattice Basis: alfa.Lattice{1,Float64}([2.0])
Domain: 2-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
 [1.0]
Codomain: 2-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
 [1.0]
Multiplier: 3-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{1}([-1], [0.0 1.0; 0.0 0.0])
 alfa.Multiplier{1}([0], [-2.0 1.0; 1.0 -2.0])
 alfa.Multiplier{1}([1], [0.0 0.0; 1.0 0.0])

alfa.wrtLattice — Method

wrtLattice(C::Crystal, L::Lattice)
wrtLattice(C::Crystal{N,T}, A) where {N,T}

Rewrites the crystal with respect to L::Lattice or lattice basis A. Thus, the lattice must be a sublattice of C.L.

Example

julia> using alfa

julia> using LinearAlgebra

julia> C = alfa.Crystal{2,Float64}()
Lattice Basis: alfa.Lattice{2,Float64}([1.0 0.0; 0.0 1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]

julia> L = alfa.Lattice{2,Float64}(2*I)
alfa.Lattice{2,Float64}([2.0 0.0; 0.0 2.0])

julia> alfa.wrtLattice(C,L)
Lattice Basis: alfa.Lattice{2,Float64}([2.0 0.0; 0.0 2.0])
Domain: 4-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
 [1.0, 0.0]
 [0.0, 1.0]
 [1.0, 1.0]
Codomain: 4-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
 [1.0, 0.0]
 [0.0, 1.0]
 [1.0, 1.0]

alfa.wrtSameLatticeAndNormalize — Method

wrtSameLatticeAndNormalize(A::CrystalOperator, B::CrystalOperator)

Finds a least common multiple translationally invariance C and rewrites both operators A and B wrt C and normalizes the Operators.

Example

julia> using alfa

julia> A = alfa.gallery.Laplace(N=1)
Lattice Basis: alfa.Lattice{1,Float64}([1.0])
Domain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Multiplier: 3-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{1}([-1], [1.0])
 alfa.Multiplier{1}([0], [-2.0])
 alfa.Multiplier{1}([1], [1.0])

julia> B = alfa.gallery.fw_restriction(N=1)
Lattice Basis: alfa.Lattice{1,Float64}([2.0])
Domain: 2-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
 [1.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Multiplier: 2-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{1}([-1], Rational{Int64}[0//1 1//2])
 alfa.Multiplier{1}([0], Rational{Int64}[1//1 1//2])

julia> (A2,B2) = alfa.wrtSameLatticeAndNormalize(A,B);

julia> A2
Lattice Basis: alfa.Lattice{1,Float64}([2.0])
Domain: 2-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
 [1.0]
Codomain: 2-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
 [1.0]
Multiplier: 3-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{1}([-1], [0.0 1.0; 0.0 0.0])
 alfa.Multiplier{1}([0], [-2.0 1.0; 1.0 -2.0])
 alfa.Multiplier{1}([1], [0.0 0.0; 1.0 0.0])

julia> B2
Lattice Basis: alfa.Lattice{1,Float64}([2.0])
Domain: 2-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
 [1.0]
Codomain: 1-element Array{StaticArrays.SArray{Tuple{1},Float64,1,1},1}:
 [0.0]
Multiplier: 2-element Array{alfa.Multiplier,1}:
 alfa.Multiplier{1}([-1], Rational{Int64}[0//1 1//2])
 alfa.Multiplier{1}([0], Rational{Int64}[1//1 1//2])