alfa.jl

alfa.Lattice

Base.getindex

Base.getproperty

Base.lcm

Base.size

alfa.CheckIfNormal

alfa.ElementsInQuotientSpace

alfa.ShiftIntoUnitCell

alfa.hnf

alfa.lll

alfa.snf_with_transform

alfa.Lattice — Method

Lattice{N,T}(A=nothing)

Construct a Lattice{N,T} with basis $A$, i.e., it represents the set

\[Aℤ^N = \left\{ ∑_{i=0}^n z_j ⋅ a_j : z_j ∈ ℤ \right\},\]

where $a_j$ denotes the $j$th column of $A$. The matrix $A$ must be square and nonsingular:

$A$ is called the lattice basis.

$a_j$ are the primite vectors of the lattice.

T<:Union{Float64,Rational} represents the datatype of the entries of the basis $a_{ij}$.

In case of A==nothing, the identity I of size N is used.

Example

julia> using alfa

julia> L = alfa.Lattice{3,Float64}()
alfa.Lattice{3,Float64}([1.0 0.0 0.0; 0.0 1.0 0.0; 0.0 0.0 1.0])

julia> L = alfa.Lattice{2,Rational{BigInt}}()
alfa.Lattice{2,Rational{BigInt}}(Rational{BigInt}[1//1 0//1; 0//1 1//1])

julia> L = alfa.Lattice{2,Rational{BigInt}}([1 2; 3 4])
alfa.Lattice{2,Rational{BigInt}}(Rational{BigInt}[1//1 2//1; 3//1 4//1])

julia> L = alfa.Lattice{2,Rational{BigInt}}([1 2; 1 2])
ERROR: AssertionError: Basis must be nonsingular

Base.getindex — Method

Base.getindex(L::Lattice, y...)

simply wrapping getindex(L.A, y...).

Example

julia> using alfa

julia> L = alfa.Lattice{2,Float64}([1 2; 3 4])
alfa.Lattice{2,Float64}([1.0 2.0; 3.0 4.0])
julia> L[1,2] == L.A[1,2]
true

Base.getproperty — Method

Base.getproperty(L::Lattice, sym::Symbol)

Get properties of Lattice{N,T}. Let $A=$L.A, i.e., $A ∈ T^{N × N}$, then

L.n and L.dim return the dimension N.

L.iA returns $A^{-1}$, i.e., the inverse of the lattice basis

L.dA returns $A^{-T}$, i.e., the basis of the dual lattice.

Example

julia> using alfa

julia> L = alfa.Lattice{2,Float64}([1 2; 3 4])
alfa.Lattice{2,Float64}([1.0 2.0; 3.0 4.0])
julia> L.n == L.dim == typeof(L).parameters[1] == 2
true
julia> L[2,1]
3.0
julia> L.iA == inv(L.A)
true
julia> L.dA == transpose(inv(L.A))
true

Base.lcm — Method

Base.lcm(A::MArray{X,T},B::MArray{X,T}...)
Base.lcm(A::MArray{X,T}, B::MArray{X,T}...) where {X,T<:Rational}

Base.lcm(A::Lattice{N,T}, B::Lattice{N,T}...) where {N,T}
Base.lcm(A::MArray{X,T}, B::MArray{X,T}...) where {X,T<:Rational}

Returns the least common multiple of $A$ and $B$ (or more), i.e. a sub-lattice $C$, that is $C ⊂ A$ and $C ⊂ B$ with $|\det(C)|$ as small as possible.

Example

julia> using alfa

julia> L = alfa.Lattice{2,Float64}([1 2; 3 4])
alfa.Lattice{2,Float64}([1.0 2.0; 3.0 4.0])

julia> alfa.lcm(L) == L
true

julia> alfa.lcm(L.A) == L.A
true

julia> alfa.lcm(alfa.Lattice{1,Float64}([2]), alfa.Lattice{1,Float64}([3]), alfa.Lattice{1,Float64}([6]))
alfa.Lattice{1,Float64}([6.0])

julia> alfa.lcm(alfa.Lattice{2,Rational{BigInt}}([1 1; -1 1]), alfa.Lattice{2,Rational{BigInt}}([1 2; 2 1]))
alfa.Lattice{2,Rational{BigInt}}(Rational{BigInt}[-3//1 1//1; -3//1 -1//1])

Base.size — Method

Base.size(L::Lattice)

Returns the dimension of the lattice basis: size(L.A,1).

Example

julia> using alfa

julia> L = alfa.Lattice{2,Float64}([1 2; 3 4])
alfa.Lattice{2,Float64}([1.0 2.0; 3.0 4.0])
julia> size(L) == size(L.A) == (2,2)
true

alfa.CheckIfNormal — Method

CheckIfNormal(s, A)

Checks if s is found within the primtive cell of A, i.e., $A^{-1}s_j ∈ [0,1)^N$ and if s is sorted lexicographically.

Example

julia> using alfa

julia> using LinearAlgebra

julia> L = alfa.Lattice{2,Float64}([1 0; 0 1]);

julia> s = [[-1/2,0], [1/4,2]]
2-element Array{Array{Float64,1},1}:
 [-0.5, 0.0]
 [0.25, 2.0]

julia>  alfa.CheckIfNormal(s,L)
false

julia> t = [[1/4,0], [1/2,0]]
2-element Array{Array{Float64,1},1}:
 [0.25, 0.0]
 [0.5, 0.0]

julia> alfa.CheckIfNormal(t,L)
true

alfa.ElementsInQuotientSpace — Method

ElementsInQuotientSpace(
    A::Union{Matrix{T},MMatrix{N,N,T}},
    B::Union{Matrix{T},MMatrix{N,N,T}};
    return_diag_hnf::Bool = false,
    return_fractional::Bool = false,
) where {N,T}

Returns all lattice points of the lattice generated by $A$ found in the primitive cell of B, i.e.,

\[T_{A,B}=\{x : x ∈ A \mathbb{Z}^N \cap B[0,1)^N \}\]

Example

julia> using alfa

julia> A = alfa.Lattice{2,Float64}([1 0; 0 1]);

julia> B = alfa.Lattice{2,Float64}([2 0; 0 2]);

julia> alfa.ElementsInQuotientSpace(A.A,B.A)
4-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
 [1.0, 0.0]
 [0.0, 1.0]
 [1.0, 1.0]

alfa.ShiftIntoUnitCell — Method

ShiftIntoUnitCell(s, A::Union{Matrix{T},MMatrix{N,N,T}}) where {N,T}
ShiftIntoUnitCell(s, A::Union{Matrix{T},MMatrix{N,N,T}}) where {N,T<:Rational}
ShiftIntoUnitCell(s, A::Lattice)

Shifts all elements s[i] into the primitive cell $A[0,1)^N$ and sort the entries lexicographically. The function returns t, y, p, such that

\[t_j + A y_j = s_{p(j)}\]

t is the shifted vector s, i.e., $A^{-1}t_j ∈ [0,1)^N$ for all j.

p is the permutation with respect to the input s.

y corresponds to the shift in fractional coordinates.

Example

julia> using alfa

julia> using LinearAlgebra

julia> L = alfa.Lattice{2,Float64}([1 0; 0 1]);

julia> s = [[-1/2,0], [1/4,2]]
2-element Array{Array{Float64,1},1}:
 [-0.5, 0.0]
 [0.25, 2.0]

julia> (t, y, p) = alfa.ShiftIntoUnitCell(s,L);

julia> t
2-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.25, 0.0]
 [0.5, 0.0]

julia> y
2-element Array{StaticArrays.MArray{Tuple{2},Float64,1,2},1}:
 [0.0, 2.0]
 [-1.0, 0.0]

julia> p
2-element Array{Int64,1}:
 2
 1

julia> [t[j] + L.A*y[j] - s[p[j]] for j in [1,2]]
2-element Array{StaticArrays.SArray{Tuple{2},Float64,1,2},1}:
 [0.0, 0.0]
 [0.0, 0.0]

alfa.hnf — Method

hnf(mat::MMatrix{M,N}) where {M, N}
hnf(mat::Matrix)

Wrapper of Nemo.hnf. Input is converted to BigInt. Returns H = mat*U, s.t. H is in Hermite Normal Form and U is unimodular.

Example

julia> using alfa # hide

julia> using LinearAlgebra # hide

julia> mat = rand(1:1000, 2, 2);

julia> H = alfa.hnf(mat);

julia> norm(LinearAlgebra.tril(H) - H) ≈ 0
true
julia> round(abs(det(inv(mat)*H)), digits=5)
1.0

alfa.lll — Method

lll(mat::MMatrix{M,N}) where {M, N}
lll(mat::Matrix)

Wrapper of Nemo.lll. Input is converted to BigInt. Applies the LLL-Algorithm to the input mat. Computes output L, such that mat*T=L for some unimodular T.

Example

julia> using alfa # hide

julia> using LinearAlgebra # hide

julia> mat = rand(1:1000, 2, 2);

julia> L = alfa.lll(mat);

julia> round(abs(det(inv(mat)*L)), digits=5)
1.0

alfa.snf_with_transform — Method

snf_with_transform(L::Lattice)
snf_with_transform(mat::MMatrix{M,N}) where {M,N}
snf_with_transform(mat::Matrix)

Wrapper of Nemo.snfwithtransform. Input is converted to BigInt. Returns (S,U,V) such that UmatV = S, where S is the Smith normal form of mat.

Example

julia> using alfa # hide

julia> using LinearAlgebra # hide

julia> mat = rand(1:10, 10, 10);

julia> (S,U,V) = alfa.snf_with_transform(mat);

julia> norm(U*mat*V - S) ≈ 0
true

julia> abs(det(U)) ≈ abs(det(V)) ≈ 1
true

julia> norm(diagm(diag(S)) - S) ≈ 0
true